+++ poc/undo/include/DeleteEdgeTransaction.hpp
@@ -11,32 +11,32 @@ class SimpleMesh;
 class DeleteEdgeTransaction : public BaseMeshTransaction<DeleteEdgeTransaction, SimpleMesh>
 {
 public:
-    static TransactionPtr create(SimpleMesh& mesh, const int id, const int start, const int end)
+    static TransactionPtr create(SimpleMesh& mesh, const size_t id, const size_t start, const size_t end)
     {
         return std::make_unique<DeleteEdgeTransaction>(mesh, id, start, end);
     }
 
-    DeleteEdgeTransaction(SimpleMesh& mesh, const int id, const int start, const int end) : BaseMeshTransaction<DeleteEdgeTransaction, SimpleMesh>(mesh), edgeId_(id), start_(start), end_(end) {}
+    DeleteEdgeTransaction(SimpleMesh& mesh, const size_t id, const size_t start, const size_t end) : BaseMeshTransaction<DeleteEdgeTransaction, SimpleMesh>(mesh), edgeId_(id), start_(start), end_(end) {}
 
-    int edgeId() const
+    size_t edgeId() const
     {
         return edgeId_;
     }
 
-    int start() const
+    size_t start() const
     {
         return start_;
     }
 
-    int end() const
+    size_t end() const
     {
         return end_;
     }
 
 private:
-    int edgeId_;
-    int start_;
-    int end_;
+    size_t edgeId_;
+    size_t start_;
+    size_t end_;
 };
 
 #endif // DELETE_NODE_TRANSACTION__HPP
